requirements.txt
customtkinter
musicbrainzngs
python-vlc
numpy
pyobjc-framework-Cocoa
discid
requests
pillow
